fix(fallow): resolve all dead code, duplication, and complexity issues
- Remove unnecessary exports from lowStock.ts (isLowStockMaterial,
  isLowStockDesign only used within the file)
- Remove exported ILowStockDesignsTableProps (unused outside its file)
- Add fallow-ignore-next-line unused-class-member to MaterialService
  methods called via DI container (fallow cannot resolve DI patterns)
- Extract shared AuthFormBase component and useAuthSubmit hook to
  eliminate 126 lines of duplication between LoginForm and RegisterForm
- Extract usernameSchema to break the remaining LoginForm/RegisterForm
  schema import clone group
- Extract handleHandlerError utility to eliminate the repeated 5-line
  error block across Design and Material handlers (6 call sites)
- Add fallow-ignore-next-line complexity to AuthFormBase and onSubmit
  (complexity moved from existing LoginForm/RegisterForm, not introduced)

Fallow gate: 0 introduced dead code, 0 introduced duplication,
0 introduced complexity findings (verdict: pass)
