+++ packages/web/src/components/LoginForm/index.tsx
@@ -1,28 +1,11 @@
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useAuth, useAuthConfig } from '@imapps/web-utils';
-import { Eye, EyeOff } from 'lucide-react';
 import type React from 'react';
-import { useState } from 'react';
-import { flushSync } from 'react-dom';
-import { useForm } from 'react-hook-form';
-import { useNavigate } from 'react-router-dom';
 import { z } from 'zod';
 
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-
-import { HOME_PAGE } from '../../constants/routes';
-import { useAlert } from '../../context/Alert';
-import { AlertStoreActions } from '../../context/Alert/types';
+import { AuthFormBase } from '@/components/AuthFormBase';
+import { usernameSchema } from '@/utils/authSchemas';
 
 const loginSchema = z.object({
-    username: z
-        .string()
-        .min(1, 'Username is required')
-        .min(3, 'Username must be at least 3 characters')
-        .max(50, 'Username must not exceed 50 characters')
-        .trim(),
+    username: usernameSchema,
     password: z
         .string()
         .min(1, 'Password is required')
@@ -30,108 +13,13 @@ const loginSchema = z.object({
         .max(100, 'Password must not exceed 100 characters'),
 });
 
-type LoginFormData = z.infer<typeof loginSchema>;
-
-export const LoginForm: React.FC = () => {
-    const {
-        handleSubmit,
-        register,
-        formState: { errors },
-    } = useForm<LoginFormData>({
-        resolver: zodResolver(loginSchema),
-    });
-    const [showPassword, setShowPassword] = useState(false);
-    const [isLoading, setIsLoading] = useState(false);
-    const { dispatch } = useAlert();
-    const navigate = useNavigate();
-    const { login } = useAuth();
-    const config = useAuthConfig();
-
-    const onSubmit = async (data: LoginFormData) => {
-        setIsLoading(true);
-        try {
-            const response = await fetch(`${config.authUrl}/login`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify(data),
-                credentials: 'include',
-            });
-
-            if (!response.ok) {
-                const json = await response.json();
-
-                throw new Error(json.message);
-            }
-
-            const json = await response.json();
-
-            if (!json.accessToken) throw new Error('No access token returned');
-            flushSync(() => login(json.accessToken));
-            navigate(HOME_PAGE.route);
-        } catch (e) {
-            const message = e instanceof Error ? e.message : 'Unknown error';
-
-            dispatch({
-                type: AlertStoreActions.SHOW_ALERT,
-                payload: {
-                    title: 'Login Error',
-                    message,
-                    severity: 'error',
-                    variant: 'standard',
-                },
-            });
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    return (
-        <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-4 w-full">
-            <div className="space-y-2">
-                <Label htmlFor="username">Username</Label>
-                <Input
-                    id="username"
-                    type="text"
-                    autoComplete="username"
-                    {...register('username')}
-                    className={errors.username ? 'border-destructive' : ''}
-                    aria-invalid={errors.username ? 'true' : 'false'}
-                />
-                <p className="text-sm text-destructive min-h-[20px]">{errors.username?.message || ''}</p>
-            </div>
-
-            <div className="space-y-2">
-                <Label htmlFor="password">Password</Label>
-                <div className="relative">
-                    <Input
-                        id="password"
-                        type={showPassword ? 'text' : 'password'}
-                        autoComplete="current-password"
-                        {...register('password')}
-                        className={`pr-10 ${errors.password ? 'border-destructive' : ''}`}
-                        aria-invalid={errors.password ? 'true' : 'false'}
-                    />
-                    <button
-                        type="button"
-                        onClick={() => setShowPassword((show) => !show)}
-                        className="absolute inset-y-0 right-0 pr-3 flex items-center"
-                        aria-label={showPassword ? 'Hide password' : 'Show password'}
-                    >
-                        {showPassword ? (
-                            <EyeOff className="h-4 w-4 text-muted-foreground" />
-                        ) : (
-                            <Eye className="h-4 w-4 text-muted-foreground" />
-                        )}
-                    </button>
-                </div>
-                <p className="text-sm text-destructive min-h-[20px]">{errors.password?.message || ''}</p>
-            </div>
-
-            <Button type="submit" disabled={isLoading} className="w-full">
-                {isLoading ? 'Logging in...' : 'Login'}
-            </Button>
-        </form>
-    );
-};
+export const LoginForm: React.FC = () => (
+    <AuthFormBase
+        schema={loginSchema}
+        endpoint="/login"
+        errorTitle="Login Error"
+        submitLabel="Login"
+        loadingLabel="Logging in..."
+        passwordAutoComplete="current-password"
+    />
+);
