+++ packages/web/src/utils/authSchemas.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const usernameSchema = z
+    .string()
+    .min(1, 'Username is required')
+    .min(3, 'Username must be at least 3 characters')
+    .max(50, 'Username must not exceed 50 characters')
+    .trim();
